Fix compiler warnings found with gcc 4.0.4.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@11534 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ test/pixmap_browser.cxx
@@ -69,7 +69,7 @@ void load_file(const char *n) {
     float fw = img->w() / float(b->w());
     float fh = img->h() / float(b->h());
     float f = fw > fh ? fw : fh;
-    b->image(img->copy(img->w()/f, img->h()/f));
+    b->image(img->copy(int(img->w()/f), int(img->h()/f)));
     img->release();
   }
 #endif
